Implement 64-bit integer printing in uart_printf()

BUG=chrome-os-partner:7490
TEST=timerinfo; numbers should look reasonable

+++ common/charge_state.h
@@ -10,7 +10,7 @@
 #define __CROS_EC_CHARGE_STATE_H
 
 /* Time constants */
-#define MSEC    ((uint64_t)1000)
+#define MSEC    (1000ULL)
 #define SECOND  (MSEC * 1000)
 #define MINUTE  (SECOND * 60)
 #define HOUR    (MINUTE * 60)
